kafka-producer/requirements.txt
requests
kafka-python
schedule